+++ src/app/cli/src/app.rs
@@ -12,6 +12,7 @@ use std::sync::Arc;
 
 use container_runtime::{ContainerRuntime, ContainerRuntimeConfig};
 use dill::*;
+use kamu::domain::compact_service::CompactService;
 use kamu::domain::*;
 use kamu::*;
 
@@ -235,6 +236,11 @@ pub fn configure_base_catalog(
     b.add::<TransformServiceImpl>();
 
     b.add::<VerificationServiceImpl>();
+
+    b.add_builder(
+        CompactServiceImpl::builder().with_run_info_dir(workspace_layout.run_info_dir.clone()),
+    );
+    b.bind::<dyn CompactService, CompactServiceImpl>();
 
     b.add::<SearchServiceImpl>();